+++ src/components/abciapp/src/abci/server/callback/mod.rs
@@ -56,10 +56,16 @@ lazy_static! {
 }
 
 #[cfg(feature = "debug_env")]
-pub const DISBALE_EVM_BLOCK_HEIGHT: i64 = 0;
+pub const DISBALE_EVM_BLOCK_HEIGHT: i64 = 1;
 
 #[cfg(not(feature = "debug_env"))]
 pub const DISBALE_EVM_BLOCK_HEIGHT: i64 = 148_3286;
+
+#[cfg(feature = "debug_env")]
+pub const ENABLE_FRC20_HEIGHT: i64 = 1;
+
+#[cfg(not(feature = "debug_env"))]
+pub const ENABLE_FRC20_HEIGHT: i64 = 149_7900;
 
 pub fn info(s: &mut ABCISubmissionServer, req: &RequestInfo) -> ResponseInfo {
     let mut resp = ResponseInfo::new();
@@ -108,7 +114,7 @@ pub fn init_chain(
 pub fn check_tx(s: &mut ABCISubmissionServer, req: &RequestCheckTx) -> ResponseCheckTx {
     let mut resp = ResponseCheckTx::new();
 
-    let tx_catalog = try_tx_catalog(req.get_tx());
+    let tx_catalog = try_tx_catalog(req.get_tx(), false);
 
     let td_height = TENDERMINT_BLOCK_HEIGHT.load(Ordering::Relaxed);
 
@@ -130,7 +136,7 @@ pub fn check_tx(s: &mut ABCISubmissionServer, req: &RequestCheckTx) -> ResponseC
             resp
         }
         TxCatalog::EvmTx => {
-            if DISBALE_EVM_BLOCK_HEIGHT < td_height {
+            if DISBALE_EVM_BLOCK_HEIGHT < td_height && td_height < ENABLE_FRC20_HEIGHT {
                 resp.code = 2;
                 resp.log = "EVM is disabled".to_owned();
                 resp
@@ -192,7 +198,7 @@ pub fn begin_block(
         pnk!(la.update_staking_simulator());
     }
 
-    if DISBALE_EVM_BLOCK_HEIGHT < header.height {
+    if DISBALE_EVM_BLOCK_HEIGHT < header.height && header.height < ENABLE_FRC20_HEIGHT {
         ResponseBeginBlock::default()
     } else {
         s.account_base_app.write().begin_block(req)
@@ -205,7 +211,7 @@ pub fn deliver_tx(
 ) -> ResponseDeliverTx {
     let mut resp = ResponseDeliverTx::new();
 
-    let tx_catalog = try_tx_catalog(req.get_tx());
+    let tx_catalog = try_tx_catalog(req.get_tx(), true);
     let td_height = TENDERMINT_BLOCK_HEIGHT.load(Ordering::Relaxed);
     const EVM_FIRST_BLOCK_HEIGHT: i64 = 142_5000;
 
@@ -234,7 +240,9 @@ pub fn deliver_tx(
                         }
                     }
 
-                    if DISBALE_EVM_BLOCK_HEIGHT < td_height {
+                    if DISBALE_EVM_BLOCK_HEIGHT < td_height
+                        && td_height < ENABLE_FRC20_HEIGHT
+                    {
                         if is_convert_account(&tx) {
                             resp.code = 2;
                             resp.log = "EVM is disabled".to_owned();
@@ -247,7 +255,7 @@ pub fn deliver_tx(
                         if let Err(err) =
                             s.account_base_app.write().deliver_findora_tx(&tx)
                         {
-                            log::debug!(target: "abciapp", "deliver convert account tx failed: {:?}", err);
+                            log::info!(target: "abciapp", "deliver convert account tx failed: {:?}", err);
 
                             resp.code = 1;
                             resp.log =
@@ -299,7 +307,7 @@ pub fn deliver_tx(
             resp
         }
         TxCatalog::EvmTx => {
-            if DISBALE_EVM_BLOCK_HEIGHT < td_height {
+            if DISBALE_EVM_BLOCK_HEIGHT < td_height && td_height < ENABLE_FRC20_HEIGHT {
                 resp.code = 2;
                 resp.log = "EVM is disabled".to_owned();
                 resp
@@ -367,7 +375,7 @@ pub fn end_block(
         &begin_block_req.byzantine_validators.as_slice(),
     );
 
-    if ! DISBALE_EVM_BLOCK_HEIGHT < td_height {
+    if td_height <= DISBALE_EVM_BLOCK_HEIGHT || td_height >= ENABLE_FRC20_HEIGHT {
         let _ = s.account_base_app.write().end_block(req);
     }
 
@@ -392,13 +400,12 @@ pub fn commit(s: &mut ABCISubmissionServer, req: &RequestCommit) -> ResponseComm
         .and_then(|s| fs::write(&path, s).c(d!(path))));
 
     let mut r = ResponseCommit::new();
+    let la_hash = state.get_state_commitment().0.as_ref().to_vec();
+    let cs_hash = s.account_base_app.write().commit(req).data;
 
-    if DISBALE_EVM_BLOCK_HEIGHT < td_height {
-        let la_hash = state.get_state_commitment().0.as_ref().to_vec();
+    if DISBALE_EVM_BLOCK_HEIGHT < td_height && td_height < ENABLE_FRC20_HEIGHT {
         r.set_data(la_hash);
     } else {
-        let la_hash = state.get_state_commitment().0.as_ref().to_vec();
-        let cs_hash = s.account_base_app.write().commit(req).data;
         r.set_data(app_hash("commit", td_height, la_hash, cs_hash));
     }
 
@@ -413,7 +420,7 @@ fn app_hash(
     mut la_hash: Vec<u8>,
     mut cs_hash: Vec<u8>,
 ) -> Vec<u8> {
-    log::debug!(target: "abciapp",
+    log::info!(target: "abciapp",
         "app_hash_{}: {}_{}, height: {}",
         when,
         hex::encode(la_hash.clone()),